peppa/requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.14.0
django-filter==23.1
django-redis==5.2.0
django-redis-cache==3.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
importlib-metadata==6.6.0
Markdown==3.4.3
mysqlclient==2.1.1
Pygments==2.15.1
PyJWT==1.7.1
PyMySQL==1.0.3
pytz==2023.3
redis==3.5.3
sqlparse==0.4.4
supervisor==4.2.5
uWSGI==2.0.21
zipp==3.15.0
